BV-34 - Add primer trimming
- Modified the maximum mismatches options for primer trimming so that the operation takes into account all matched bases as opposed to only those that are specified by the intervals of the Smith-Waterman alignment.
- Added tests for the maximum mismatches and minimum match length options for primer trimming.

+++ test/com/biomatters/plugins/biocode/labbench/lims/LimsTestCase.java
@@ -100,7 +100,7 @@ public abstract class LimsTestCase extends Assert {
             PlateDocument plateDoc = (PlateDocument) doc.getDocumentOrNull();
             if(plateDoc.getPlate().getReactionType() == expectedType && plateName.equals(doc.getName())) {
                 Reaction[] reactions = plateDoc.getPlate().getReactions();
-                assertEquals(Plate.Size.w96.numberOfReactions(), reactions.length);
+                assertEquals((int)Plate.Size.w96.numberOfReactions(), reactions.length);
                 for(int i=0; i<expectedExtractionIds.length; i++) {
                     assertEquals(expectedExtractionIds[i], reactions[i].getExtractionId());
                 }
